web-dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
kubernetes==31.0.0
pyyaml==6.0.2
python-jose[cryptography]==3.4.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
